requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
langchain
langchain-community
langchain-openai
openai>=1.3.0
tiktoken>=0.5.1
python-jose>=3.3.0
passlib>=1.7.4
httpx>=0.25.1